+++ libs/libc/pthread/pthread_condbroadcast.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * sched/pthread/pthread_mutexinconsistent.c
+ * libs/libc/pthread/pthread_condbroadcast.c
  *
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -28,62 +28,66 @@
 
 #include <pthread.h>
 #include <sched.h>
-#include <assert.h>
 #include <errno.h>
+#include <debug.h>
 
-#include <nuttx/sched.h>
-#include <nuttx/semaphore.h>
-#include <nuttx/pthread.h>
+#include <nuttx/atomic.h>
 
-#include "pthread/pthread.h"
-#include "sched/sched.h"
+#include "pthread.h"
 
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
 
 /****************************************************************************
- * Name: pthread_mutex_inconsistent
+ * Name: pthread_cond_broadcast
  *
  * Description:
- *   This function is called when a pthread is terminated via either
- *   pthread_exit() or pthread_cancel().  It will check for any mutexes
- *   held by exiting thread.  It will mark them as inconsistent and
- *   then wake up the highest priority waiter for the mutex.  That
- *   instance of pthread_mutex_lock() will then return EOWNERDEAD.
+ *    A thread broadcast on a condition variable.
  *
  * Input Parameters:
- *   tcb -- a reference to the TCB of the exiting pthread.
+ *   None
  *
  * Returned Value:
- *   None.
+ *   None
+ *
+ * Assumptions:
  *
  ****************************************************************************/
 
-void pthread_mutex_inconsistent(FAR struct tcb_s *tcb)
+int pthread_cond_broadcast(FAR pthread_cond_t *cond)
 {
-  FAR struct pthread_mutex_s *mutex;
-  FAR struct tls_info_s *tls = nxsched_get_tls(tcb);
+  int ret = OK;
 
-  DEBUGASSERT(tcb != NULL);
+  sinfo("cond=%p\n", cond);
 
-  nxmutex_lock(&tls->tl_lock);
-
-  /* Remove and process each mutex held by this task */
-
-  while (tls->tl_mhead != NULL)
+  if (!cond)
     {
-      /* Remove the mutex from the TCB list */
+      ret = EINVAL;
+    }
+  else
+    {
+      int wcnt = atomic_read(COND_WAIT_COUNT(cond));
 
-      mutex         = tls->tl_mhead;
-      tls->tl_mhead = mutex->flink;
-      mutex->flink  = NULL;
+      /* Loop until all of the waiting threads have been restarted. */
 
-      /* Mark the mutex as INCONSISTENT and wake up any waiting thread */
+      while (wcnt > 0)
+        {
+          if (atomic_cmpxchg(COND_WAIT_COUNT(cond), &wcnt, wcnt - 1))
+            {
+              /* Post the condition semaphore to wake up a waiting thread.
+               * Only the highest priority waiting thread will get to execute
+               */
 
-      mutex->flags |= _PTHREAD_MFLAGS_INCONSISTENT;
-      mutex_reset(&mutex->mutex);
+              ret = -nxsem_post(&cond->sem);
+
+              /* Decrement the waiter count */
+
+              wcnt--;
+            }
+        }
     }
 
-  nxmutex_unlock(&tls->tl_lock);
+  sinfo("Returning %d\n", ret);
+  return ret;
 }
