+++ libs/libc/pthread/pthread_mutex.c
@@ -36,6 +36,8 @@
 #include <nuttx/semaphore.h>
 #include <nuttx/pthread.h>
 
+#include "pthread.h"
+
 /****************************************************************************
  * Private Functions
  ****************************************************************************/
