+++ libs/libc/pthread/pthread_mutex_consistent.c
@@ -35,6 +35,8 @@
 #include <nuttx/semaphore.h>
 #include <nuttx/pthread.h>
 
+#include "pthread.h"
+
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
