+++ libs/libc/pthread/pthread_mutex_init.c
@@ -34,6 +34,8 @@
 #include <nuttx/semaphore.h>
 #include <nuttx/pthread.h>
 
+#include "pthread.h"
+
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
