+++ libs/libc/pthread/pthread_mutex_timedlock.c
@@ -36,6 +36,8 @@
 #include <nuttx/sched.h>
 #include <nuttx/pthread.h>
 
+#include "pthread.h"
+
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
