+++ libs/libc/pthread/pthread_mutex_trylock.c
@@ -35,6 +35,8 @@
 
 #include <nuttx/pthread.h>
 
+#include "pthread.h"
+
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
